+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @enschedule/dashboard
 
+## 1.1.17
+
+### Patch Changes
+
+- Fix runNow type and poll for logs when schedule is running
+- Updated dependencies
+  - @enschedule/types@1.1.6
+  - @enschedule/worker@1.1.12
+  - @enschedule/worker-api@1.1.6
+
 ## 1.1.16
 
 ### Patch Changes