+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @enschedule/cli
 
+## 1.1.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @enschedule/types@1.1.6
+  - @enschedule/worker-api@1.1.6
+
 ## 1.1.10
 
 ### Patch Changes